Complete version of Wasteagram
some of the features do not work anymore since the license of the Google Firestore database I used ran out, you can get the gist of the app. There is a timeline of all the photos uploaded with the number of wasted items. You would have been able to add a post by clicking the blue button on the botttom, taking a picture, entering the number of waste items that are in the pictures, and pressing the blue upload button. The timeline would have shown the total number of wasted items from all the posts.